メインプロセスで画像ファイルパスの加工処理を動作させるようにした
preload.jsで特定のモジュール以外のモジュールを埋め込みできなくなったため
https://www.electronjs.org/docs/latest/tutorial/sandbox
